+++ src/app/(route)/_landing/_components/CooperationSection/CooperationSection.tsx
