+++ src/app/(route)/_landing/_components/DetailSection/DetailSection.stories.tsx
