+++ src/app/(route)/_landing/_components/HeroSection/HeroSection.stories.tsx
