+++ src/app/(route)/_landing/_components/KanbanSection/KanbanSection.tsx
