+++ src/app/(route)/_landing/_components/_internal/ResponsiveImage/ResponsiveImage.tsx
