+++ src/app/(route)/_landing/_constants/index.ts
